Move get random functions into a util module
