+++ discord-bot/src/utils/random.ts
@@ -1,0 +1,10 @@
+export const getRandomIntInclusive = (min: number, max: number) => {
+  const intMin = Math.ceil(min);
+  const intMax = Math.floor(max);
+  return Math.floor(Math.random() * (intMax - intMin + 1) + min);
+};
+
+export const getRandomBoolean = (): boolean => {
+  const randomBinaryNumber = Math.floor(Math.random() * Math.floor(2));
+  return Boolean(randomBinaryNumber);
+};
